+++ precompiles/outposts/stride/setup_test.go
@@ -6,6 +6,9 @@ package stride_test
 import (
 	"testing"
 
+	"github.com/evmos/evmos/v16/precompiles/erc20"
+
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/evmos/evmos/v16/precompiles/outposts/stride"
 	"github.com/evmos/evmos/v16/testutil/integration/common/grpc"
 	testkeyring "github.com/evmos/evmos/v16/testutil/integration/evmos/keyring"
@@ -36,6 +39,7 @@ func (s *PrecompileTestSuite) SetupTest() {
 	)
 
 	precompile, err := stride.NewPrecompile(
+		common.HexToAddress(erc20.WEVMOSContractTestnet),
 		network.App.TransferKeeper,
 		network.App.Erc20Keeper,
 		network.App.AuthzKeeper,