+++ tests/nix_tests/test_osmosis_outpost.py
@@ -1,6 +1,7 @@
 import json
 
 import pytest
+from web3 import Web3
 
 from .ibc_utils import (
     EVMOS_IBC_DENOM,
@@ -15,6 +16,7 @@ from .utils import (
     KEYS,
     OSMOSIS_POOLS,
     WASM_CONTRACTS,
+    WEVMOS_ADDRESS,
     approve_proposal,
     erc20_balance,
     eth_to_bech32,
@@ -24,8 +26,12 @@ from .utils import (
     send_transaction,
     wait_for_cosmos_tx_receipt,
     wait_for_fn,
-    wrap_evmos,
 )
+
+# This contract address is provided in genesis
+# as registered token pair. If need to edit this
+# do so in 'osmosis-outpost.jsonnet' file
+WOSMO_ADDRESS = Web3.toChecksumAddress("0x5db67696C3c088DfBf588d3dd849f44266ff0ffa")
 
 
 @pytest.fixture(scope="module", params=["evmos"])
@@ -57,15 +63,8 @@ def test_osmosis_swap(ibc):
 
     xcs_contract = setup_osmos_chains(ibc)
 
-    # --------- Register Evmos token (this could be wrapevmos I think)
-    wevmos_addr = wrap_evmos(ibc.chains["evmos"], evmos_addr, amt)
-
     # --------- Transfer Osmo to Evmos
     transfer_osmo_to_evmos(ibc, osmosis_addr, evmos_addr)
-
-    # --------- Register Osmosis ERC20 token
-    osmo_erc20_addr = register_osmo_token(evmos)
-    print(f"osmo_erc20_addr: {osmo_erc20_addr}")
 
     # define TWAP parameters
     testSlippagePercentage = 20
@@ -79,8 +78,8 @@ def test_osmosis_swap(ibc):
         "channelID": "channel-0",
         "xcsContract": xcs_contract,
         "sender": evmos_addr,
-        "input": wevmos_addr,
-        "output": osmo_erc20_addr,
+        "input": WEVMOS_ADDRESS,
+        "output": WOSMO_ADDRESS,
         "amount": amt,
         "slippagePercentage": testSlippagePercentage,
         "windowSeconds": testWindowSeconds,
@@ -97,18 +96,19 @@ def test_osmosis_swap(ibc):
 
     # check balance increase after swap
     new_erc20_balance = 0
+    # the account has 200 uosmo transferred in the setup
+    # function transfer_osmo_to_evmos
+    initial_erc20_balance = 200
 
     def check_erc20_balance_change():
         nonlocal new_erc20_balance
-        new_erc20_balance = erc20_balance(w3, osmo_erc20_addr, evmos_addr)
+        new_erc20_balance = erc20_balance(w3, WOSMO_ADDRESS, evmos_addr)
         print(f"uosmo erc20 balance: {new_erc20_balance}")
-        return new_erc20_balance > 0
+        return new_erc20_balance > initial_erc20_balance
 
     wait_for_fn("balance change", check_erc20_balance_change)
 
-    # the account has 200 uosmo IBC coins from the setup
-    # previous to registering the uosmo token pair
-    exp_final_balance = 200 + exp_swap_amount
+    exp_final_balance = initial_erc20_balance + exp_swap_amount
     assert new_erc20_balance == exp_final_balance
 
 
@@ -234,9 +234,12 @@ def transfer_osmo_to_evmos(ibc, src_addr, dst_addr):
 
     new_dst_balance = 0
 
+    # Osmo is registered as token pair since genesis.
+    # Check the ERC20 contract balance
     def check_balance_change():
         nonlocal new_dst_balance
-        new_dst_balance = get_balance(dst_chain, bech_dst, OSMO_IBC_DENOM)
+        new_dst_balance = erc20_balance(dst_chain.w3, WOSMO_ADDRESS, dst_addr)
+        print(f"uosmo erc20 balance: {new_dst_balance}")
         return old_dst_balance != new_dst_balance
 
     wait_for_fn("balance change", check_balance_change)
@@ -281,11 +284,13 @@ def register_osmo_token(evmos):
     print("proposal id: ", proposal_id)
     # vote 'yes' on proposal and wait it to pass
     approve_proposal(evmos, proposal_id)
-    # query token pairs and get WEVMOS address
+    # query token pairs and get contract address
     pairs = evmos_cli.get_token_pairs()
-    assert len(pairs) == 2
-    assert pairs[1]["denom"] == osmos_ibc_denom
-    return pairs[1]["erc20_address"]
+    assert len(pairs) == 1, "expected exactly one token pair to be registered"
+    assert (
+        pairs[0]["denom"] == osmos_ibc_denom
+    ), "expected the registered token pair to be the Osmosis IBC coin"
+    return pairs[0]["erc20_address"]
 
 
 def deploy_wasm_contract(osmosis_cli, deployer_addr, contract_file, init_args, label):